Project_1/requirements.txt
pandas==1.5.3
mysql-connector-python==8.0.33
apache-airflow==2.6.3